Move profilers into a separate module
